Handling the file operation's edge case

+++ src/migrate/command.py
@@ -1,25 +1,41 @@
+# package imports 
 import click
 import yaml
 import os
 
+# local imports 
+from .dataYAML import dumping_data
+
+
+ALLOWED_DRIVERS = [
+                'mysql', 
+                'sqlite', 
+                'postgresql', 
+                'oracle'
+            ]
+
 @click.command()
+@click.option('--driver','-d', default="mysql", help='Driver for ORM ')
 @click.version_option(version="0.1.0")
-def cli():
-    dumping_data = {
-        "db":{
-            "DRIVER":"adaptor",
-            "DB_NAME":"richorm",
-            "DB_HOST":"localhost",
-            "DB_PORT":3301,
-            "DB_PASSWORD":"localhost@password"
-        },
-        "models":{
-            "directory":"path/to/models/directory"
-        }
-    }
+def cli( driver:str ):
     
+    if driver not in ALLOWED_DRIVERS:
+        raise ValueError(
+            "This driver is not configured with richORM"
+            )
+        
+    dumping_data["database"]["driver"] = f"{driver}"
     currentDirectory  = os.getcwd()
-    with open(f'{currentDirectory}/richorm-config.yml', "w") as configFile:
-        yaml.dump(dumping_data, configFile, default_flow_style=False )
+    
+    if not os.path.exists(f'{currentDirectory}/richorm-config.yaml'):
+        with open(f'{currentDirectory}/richorm-config.yaml', "w") as configFile:
+            yaml.dump(dumping_data, configFile, default_flow_style=False )
+            
+    else:
+        raise FileExistsError(
+            "This file already exists in your directory .Please modify that file."
+        )
+    
+
